Refactor: 공지뷰 리팩토링 및 공지뷰모델 생성
