+++ PJ3T3_Postie/Core/Setting/View/NoticeView.swift
@@ -9,8 +9,7 @@ import SwiftUI
 
 struct NoticeView: View {
     @ObservedObject var firestoreNoticeManager = FirestoreNoticeManager.shared
-    @AppStorage("isThemeGroupButton") private var isThemeGroupButton: Int = 0
-    @State private var isExpanded = false
+    @StateObject private var viewModel = NoticeViewModel()
     
     var body: some View {
         ZStack {
@@ -25,7 +24,7 @@ struct NoticeView: View {
                     
                     if firestoreNoticeManager.notices.isEmpty {
                         VStack {
-                            Image(isThemeGroupButton == 4 ? "postyThinkingSketchWhite" : "postyThinkingSketch")
+                            Image(viewModel.isThemeGroupButton == 4 ? "postyThinkingSketchWhite" : "postyThinkingSketch")
                                 .resizable()
                                 .scaledToFit()
                                 .frame(height: 200)
@@ -59,7 +58,7 @@ struct NoticeView: View {
 //                                            .scaledToFit()
 //                                    }
                                     
-                                    Text("\(parseText(notice.content.replacingOccurrences(of: "\\n", with: "\n")))\n")
+                                    Text("\(viewModel.parseText(notice.content.replacingOccurrences(of: "\\n", with: "\n")))\n")
                                         .font(.callout)
                                     
                                     HStack {
@@ -107,27 +106,6 @@ struct NoticeView: View {
         .toolbarBackground(postieColors.backGroundColor, for: .navigationBar)
         .navigationBarTitleDisplayMode(.inline)
     }
-}
-
-func parseText(_ text: String) -> Text {
-    var resultText = Text("")
-    var tempText = text
-    
-    while let range = tempText.range(of: "bold(") {
-        let textBeforeBold = tempText[..<range.lowerBound]
-        tempText.removeSubrange(..<range.upperBound)
-        
-        if let endRange = tempText.range(of: ")") {
-            let boldText = tempText[..<endRange.lowerBound]
-            tempText.removeSubrange(..<endRange.upperBound)
-            
-            resultText = resultText + Text(textBeforeBold) + Text(boldText).bold()
-        }
-    }
-    
-    resultText = resultText + Text(tempText) // 나머지 텍스트 추가
-    
-    return resultText
 }
 
 //#Preview {
